add testing check

+++ connect/common/signals.py
@@ -114,12 +114,13 @@ def org_authorizations(sender, instance, created, **kwargs):
 def delete_authorizations(instance, **kwargs):
     for project in instance.organization.project.all():
         project.project_authorizations.filter(user__email=instance.user.email).delete()
-
-    ai_client = IntelligenceRESTClient()
-    ai_client.delete_user_permission(
-        organization_id=instance.organization.inteligence_organization, 
-        user_email=instance.user.email
-    )
+    
+    if not settings.TESTING:
+        ai_client = IntelligenceRESTClient()
+        ai_client.delete_user_permission(
+            organization_id=instance.organization.inteligence_organization, 
+            user_email=instance.user.email
+        )
 
     instance.organization.send_email_remove_permission_organization(
         first_name=instance.user.first_name, email=instance.user.email
